style: extract menu dropdown to its own component
